Implement Day 1 in Kotlin
I think I will try to write my finalized Kotlin solutions concise this year. I believe that trying to optimize Kotlin code vs. Java code is simply a lost cause because Kotlin simply does more things in the bytecode that Java doesn't. Kotlin is a higher-level JVM language than Java, and it shows with the worse performance. I think it is a fun challenge to try and optimize the Java code for speed while optimizing the Kotlin code for (relative) conciseness.