Solve Day 8 in Kotlin
I had a few off-by-ones in my code that meant I didn't realize the loop size for Part 2 matched the number of steps it takes to get to Z the first time.
My observations that helped me find the solution to Part 2:
1. Each A node leads to exactly one Z node
2. Each Z node is found on the very last instruction of the instruction list
3. The loop size always equals the number of steps it takes to get to Z for the first time. Specifically, the Z node points to a node which then points to the 3rd visited node including the start node. The first step count to Z is equal to Z's index. The loop size is equal to (zIdx + 2 - 2) since the size of the visited list is 2 more than the index, and the index of the start of the loop is i = 2 or the 3rd element.

After finding that the loop size and initial step count were the same, it was trivial to find the LCM for Part 2. Part 1 is just a subset of Part 2, the loop size from AAA to ZZZ.