Solve Day 10 in Kotlin
I finally placed on the leaderboard today! I got #92 on Part 1 and didn't place for Part 2.
I have basically given up on using Java as I don't actually care to optimize my solutions because I don't really have the mental energy to write a solution in 2 languages.
Using Kotlin has sped me up quite a bit because there are many helper functions, and I can write my own extension functions when I need to.
For Part 1, I didn't actually realize that the loop was continuous (even though it's called a loop).
So, I went for a Dijkstra solution for Part 1, which I still somehow ended up placing #92 with.
That solution has already been removed from the code I am pushing here.
For Part 2, I really had no idea how to approach it. I was staring at the code for the better part of an hour and a half without really knowing what I was doing.
Finally, I'm not even sure how, but I realized that I could double the grid on both dimensions and do a flood fill, translating the coordinates back to the original grid, and using that to find the inside spots.
I explained my strategy in more detail as a comment inside Day10.kt.
I got #2245 for Part 2, well outside the leaderboard.

Honestly, I had hoped to do better this year than last in terms of placement.
I'm not sure whether the problems have been harder, I've been slower at them, or the competition has been more intense. Probably a little bit of all three. I used to think my skill in sometimes placing didn't come to writing the code fast, but understanding and creating a solution quickly.
Perhaps that's still somewhat true, but my performance has just been a bit disappointing to me. Last year, I was so close to getting on the global leaderboard of all days, even if only for a moment. This year, that doesn't seem like a possibility.
I have to wonder for how long Eric will keep running Advent of Code. This is his 9th year doing it already! I can't imagine its sustainable forever.
I hope to do better in future years. I think a lot of it comes down to practice. I'm not really sure how to practice nor if I have the energy to do so between now and the next AOC (if it is happening, of course).
Anyways, I'm happy to have placed, even if small. I hope to place another few times this year if possible.