Implement Day 16 in Kotlin
Today feels a bit sad for me. I wrote Part 1 insanely fast (7 minutes), and it worked on the test input. However, it failed on my input. This was due a to a single bug where my code didn't work if there was a mirror at (0,0), which the example input didn't have. I placed #76 on Part 1.
If not for this bug, I could've placed anywhere from #10 to #21 (the people who placed in the 7th minute).
On Part 2, I had no bugs, but I also didn't go very fast, so I placed #97.
I wish I could've spotted the bug faster or realized it was possible.
Additionally, my time delta from Part 1 to Part 2 was 4 minutes and 38 seconds. Assuming I was in the later half of the 7th minute (which I'm pretty sure I was), then that would have put me in the ballpark #40 range for Part 2.
Alas! I must stop dwelling on this now. It just stinks to go so faster and get tripped up by something so small / easy to fix. I guess this is just another example of why I need more practice. :)