Implement Day 21 in Kotlin
Today's puzzle was unfortunately the only one so far that took me more than 24 hours to complete. I got Part 1 in 12 minutes and 6 seconds for a placement of #1063. For Part 2, it took me around 2 days to fully complete. I pretty early on realized that I would need to do some math to calculate the extent of the reachable plots.

From memory, I think I came upon the parts of the solution in the following order:
* I realized that there was a core set of plots where all squares were reachable
* I realized that there was a distinction between what I call "odd squares" and "even squares" -- the odd square includes the main square, and then the odd/even squares alternately tile the infinite grid
* I realized that there were fringe squares that could only partly be reached into

What took me the longest was actually figuring out the math. A HUGE bug of mine was mistaking the fact that the number of steps required to reach the edge is lost forever. I was accidentally using that step length twice. When I realized this, it actually completely changed the fringe squares, and I had to start my math for that over.
Another big revelation was that the real input was actually easier than the sample input. There is a direct line to each edge from the starting point, and the starting point is in the exact center of the main square. This simplifies things a lot.
This made me realize that the topmost square was an exact triangle shape with the upper 2 corners not reachable, with similar logic for the bottommost, leftmost, and rightmost squares.
There are 2 other types of fringe squares under this logic. For each set of northwest, northeast, southeast, southwest directions, there are a set of squares where all parts are reachable except the furthest corner, and then a similar inverted one where only the closest corner is reachable.

After realizing all this, it just becomes a challenge of exactly calculating the number of reachable plots. Since we want all garden plots that are reachable in exactly the number of steps provided, we want all plots up to and including the furthest-reachable plots but with a condition of matching "oddness," as I call it. If our required step count is odd, then the plots that we want will have an odd distance, and similar for an even step count. For the step count provided in Part 2, it is a very nice number. After subtracting out the 65 steps necessary to get to the edge in any direction, we can reach up to exactly 202,300 squares in any direction, not including the main square.

What tripped me up the most, besides double counting the initial 65 steps to the edge, was the "oddness" we wanted to calculate for specific fringe squares. It turns out I had the right strategy ages ago, but I had calculated the wrong required oddness for the fringe squares.

All in all, the puzzle was pretty interesting, but the math was annoying. I'm almost sure that I overcomplicated this, somehow. Or maybe not? My code is pretty slow, though. It takes about a second to run, but we are also talking about Kotlin here. I could definitely speed it up if I wanted to, as I calculate distances past what I actually want.

To be honest, I'm just happy that I was able to complete this puzzle without any outside help. The example input was quite annoying because it did not match the ease of the real input at all. And none of the provided sample step counts were odd, which made it effectively useless for me to be able to use that data in my simplified solution.