Implement Day 22 in Java
My solution in Kotlin for Part 2 was so bad and unsalvageable that I had to start over and switch to Java. I still don't even know what was wrong with my Kotlin solution, but it was underestimating by about 2000 bricks for Part 2. I went for an entirely different strategy in Java, so that I only had to simulate the bricks falling the first time, then build up a map of a brick to the bricks it supports and a map of a brick to the bricks supporting it using BitSets. Even after optimizing with BitSets, it still takes about a second to run. This is kind of surprising to me. I'm sure that there are faster strategies, which I may want to try and investigate in the future.

In Kotlin, I attempted to simulate the bricks falling after removing each one individually. This somehow underestimated the number, and it was also insanely slow, requiring about 40 seconds to run on my hardware.

Well, the puzzle is over. I wish I knew what was so wrong about my Kotlin code, but I don't think I care enough to find out.