Implement Day 23 in Kotlin
I placed #60 on Part 1 today! For Part 2, I didn't place and got #445. Today was not too hard. It definitely helped that I was able to copy my search() function from Day 21.

After realizing that all paths are single-width corridors with some forks in the road, I was able to condense the grid to a weighted graph, which made it fast enough to compute the longest path on my computer. It takes about 20 seconds to run both parts on my computer. Additionally, the slopes on the grid don't even matter. They are always placed right before/after a junction point, so you can just guess where they are and go off of that for Part 1.